Moved database connects outside of models directory
